feat(ui): remove AutoBePlaygroundChatSideMovie and implement useEscapeKey hook
- Deleted AutoBePlaygroundChatSideMovie component to streamline the codebase.
- Introduced useEscapeKey hook for handling ESC key press to close modals or overlays.
- Updated AutoBeStatusModal to utilize the new useEscapeKey hook for improved user experience.
