+++ packages/ui/src/hooks/index.ts
@@ -1,2 +1,3 @@
+export * from "./useEscapeKey";
 export * from "./useIsomorphicLayoutEffect";
 export * from "./useMediaQuery";
